Move EventSummary to a seperate file
